chore : Graph View Detached from statistics panel
