+++ src/Components/RoutinePage/OperationPanel/StatisticsPanel.jsx
@@ -2,8 +2,9 @@ import React, { useEffect, useRef, useState } from "react";
 import '../../../Styles/RoutinePage/OperationPanel/StatisticsPanel.css'
 import OptionButton from "./OptionButton";
 import '../../../../node_modules/react-vis/dist/style.css';
-import { FlexibleXYPlot, VerticalBarSeries, LabelSeries, XAxis, YAxis, LineSeries } from 'react-vis';
+
 import testRoutineData from "../../../Assets/testRoutineData";
+import GraphView from "./GraphView";
 
 const StatisticsPanel = (props) => {
     const graphViewRef = useRef();
@@ -16,53 +17,12 @@ const StatisticsPanel = (props) => {
         })
     });
 
-    const executionTimePublish = testRoutineData.routine.filter((data) => {
-        if (data.operation === "PUBLISH"){
-            return data;
-        }
-    }).map((data) => {
-        return {x : data.title, y : data.executionTime}
-    }) ;
-
-    const executionTimeSubscribe = testRoutineData.routine.filter((data) => {
-        if (data.operation !== "PUBLISH"){
-            return data;
-        }
-    }).map((data) => {
-        return {x : data.title, y : data.executionTime}
-    });
-
-    const [execView, updateExecView] = useState(true);
-
-    const dataByteDataPublish = testRoutineData.routine.filter((dat) => {
-        if (dat.operation === "PUBLISH"){
-            return dat;
-        }
-    }).map((dat) => {
-        return {
-            x: dat.title,
-            y: dat.dataBytes
-        }
-    })
-
-    const dataByteDataSubscribe = testRoutineData.routine.filter((dat) => {
-        if (dat.operation !== "PUBLISH"){
-            return dat;
-        }
-    }).map((dat) => {
-        return {
-            x: dat.title,
-            y: dat.dataBytes
-        }
-    })
-
     const [selectedIndex, updateSelectedIndex] = useState(null);
 
     const [Options, UpdateOptions] = useState([{ key: 0, title: "Routine Map", isSelected: false }, { key: 1, title: "Graph", isSelected: false }, { key: 2, title: "Create Routine", isSelected: false }]);
 
 
     const handleOptionClickCallback = (index) => {
-        console.log(executionTimePublish);
         const clone = [...Options];
         if (selectedIndex !== null) {
             clone[selectedIndex].isSelected = false;
@@ -76,16 +36,6 @@ const StatisticsPanel = (props) => {
         return Options.map((option) => (<OptionButton onClickOption={handleOptionClickCallback} index={option.key} isSelected={option.isSelected} title={option.title} />));
     }
 
-    const handleGraphViewOption = (index) => {
-       
-        if (index === 0) {
-            updateExecView(false);
-        }
-        else {
-            updateExecView(true);
-        }
-    }
-
     const renderPresentationComponent = () => {
         if (selectedIndex === null) {
             return <div className="warningView">
@@ -96,22 +46,7 @@ const StatisticsPanel = (props) => {
         else {
             if (selectedIndex === 1) {
                 return (
-                    <div className="graphView">
-                        <div className="graph">
-                            <FlexibleXYPlot className="plot" xType="ordinal" height={graphViewRef.current.offsetHeight * 0.7} width = {graphViewRef.current.offsetWidth * 0.75}>
-                                <XAxis/>
-                                <YAxis/>
-                                { execView === false ? <VerticalBarSeries data={dataByteDataPublish} barWidth={0.4} opacity={1} color={"rgb(245, 61, 61)"} /> : <></> } 
-                                { execView === false ? <VerticalBarSeries data={dataByteDataSubscribe} barWidth={0.4} opacity={1} color={"rgb(0, 160, 198)"} /> : <></> } 
-                                {execView === true ? <LineSeries data={executionTimePublish} color={"rgb(245, 61, 61)"}/> : <></>}
-                                {execView === true ? <LineSeries data={executionTimeSubscribe} color={"rgb(0, 160, 198)"} /> : <></>}
-                            </FlexibleXYPlot>
-                        </div>
-                        <div className="options">
-                            <OptionButton onClickOption={handleGraphViewOption} index={0} isSelected={!execView} title={"Data Exchange"} />
-                            <OptionButton onClickOption={handleGraphViewOption} index={1} isSelected={execView} title={"Execution Time"} />
-                        </div>
-                    </div>
+                    <GraphView routineData={testRoutineData} graphViewRef={graphViewRef}/>
 
                 )
             }
